+++ truffle/src/com.oracle.truffle.object/src/com/oracle/truffle/object/PropertyMap.java
@@ -4,7 +4,9 @@
  *
  * This code is free software; you can redistribute it and/or modify it
  * under the terms of the GNU General Public License version 2 only, as
- * published by the Free Software Foundation.
+ * published by the Free Software Foundation.  Oracle designates this
+ * particular file as subject to the "Classpath" exception as provided
+ * by Oracle in the LICENSE file that accompanied this code.
  *
  * This code is distributed in the hope that it will be useful, but WITHOUT
  * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
@@ -22,9 +24,10 @@
  */
 package com.oracle.truffle.object;
 
-import com.oracle.truffle.api.object.Property;
 import java.util.Iterator;
 import java.util.Map;
+
+import com.oracle.truffle.api.object.Property;
 
 /**
  * Immutable property map.
@@ -70,12 +73,6 @@ public abstract class PropertyMap implements ImmutableMap<Object, Property> {
 
     /** @since 0.17 or earlier */
     public abstract PropertyMap removeCopy(Property value);
-
-    /** @since 0.17 or earlier */
-    @Deprecated
-    public PropertyMap getParentMap() {
-        throw new UnsupportedOperationException();
-    }
 
     /** @since 0.17 or earlier */
     @Override